# board-service/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateCustomImportanceRequest:
    properties:
      color:
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
      project_id:
        type: string
    required:
    - name
    - project_id
    type: object
...
  dto.CustomImportanceResponse:
    properties:
      color:
        type: string
      createdAt:
        type: string
      displayOrder:
        type: integer
      importance_id:
        type: string
      isSystemDefault:
        type: boolean
      name:
        type: string
      project_id:
        type: string
      updatedAt:
        type: string
    type: object
  dto.CustomRoleResponse:
    properties:
      color:
        type: string
      createdAt:
        type: string
      displayOrder:
        type: integer
      isSystemDefault:
        type: boolean
      name:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
      request_id:
        type: string
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      request_id:
        type: string
    type: object
...
  dto.UpdateCustomImportanceOrderRequest:
    properties:
      importanceOrders:
        items:
          $ref: '#/definitions/dto.ImportanceOrder'
        minItems: 1
        type: array
    required:
    - importanceOrders
    type: object
  dto.UpdateCustomImportanceRequest:
    properties:
      color:
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
    type: object
  dto.UpdateCustomRoleOrderRequest:
    properties:
      roleOrders:
        items:
          $ref: '#/definitions/dto.RoleOrder'
        minItems: 1
        type: array
    required:
    - roleOrders
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@wealist.com
    name: API Support
  description: Board management API for weAlist project management platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Board Service API
  version: "1.0"
paths:
...
  /api/custom-fields/importance:
    post:
      consumes:
      - application/json
      description: Create a new custom importance for a project (ADMIN+ only)
      parameters:
      - description: Importance details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCustomImportanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CustomImportanceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create custom importance
      tags:
      - custom-fields
  /api/custom-fields/importance/{importance_id}:
    delete:
      consumes:
      - application/json
      description: Delete a custom importance (ADMIN+ only, system defaults cannot
        be deleted)
      parameters:
      - description: Importance ID
        in: path
        name: importance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete custom importance
      tags:
      - custom-fields
    get:
      consumes:
      - application/json
      description: Get a specific custom importance by ID
      parameters:
      - description: Importance ID
        in: path
        name: importance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CustomImportanceResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get custom importance
      tags:
      - custom-fields
    put:
      consumes:
      - application/json
      description: Update a custom importance (ADMIN+ only, system defaults cannot
        be updated)
      parameters:
      - description: Importance ID
        in: path
        name: importance_id
        required: true
        type: string
      - description: Importance updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCustomImportanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CustomImportanceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update custom importance
      tags:
      - custom-fields
  /api/custom-fields/projects/{project_id}/importance:
    get:
      consumes:
      - application/json
      description: Get all custom importances for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CustomImportanceResponse'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get custom importances
      tags:
      - custom-fields
  /api/custom-fields/projects/{project_id}/importance/order:
    put:
      consumes:
      - application/json
      description: Update the display order of custom importances (ADMIN+ only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Importance orders
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCustomImportanceOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update importance display order
      tags:
      - custom-fields
  /api/custom-fields/projects/{project_id}/roles:
    get:
      consumes:
      - application/json
      description: Get all custom roles for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CustomRoleResponse'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get custom roles
      tags:
      - custom-fields
  /api/custom-fields/projects/{project_id}/roles/order:
    put:
      consumes:
      - application/json
      description: Update the display order of custom roles (ADMIN+ only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Role orders
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCustomRoleOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update role display order
      tags:
      - custom-fields
  /api/custom-fields/projects/{project_id}/stages:
    get:
      consumes:
      - application/json
      description: Get all custom stages for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CustomStageResponse'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get custom stages
      tags:
      - custom-fields
  /api/custom-fields/projects/{project_id}/stages/order:
    put:
      consumes:
      - application/json
      description: Update the display order of custom stages (ADMIN+ only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Stage orders
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCustomStageOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update stage display order
      tags:
      - custom-fields
  /api/custom-fields/roles:
    post:
      consumes:
      - application/json
      description: Create a new custom role for a project (ADMIN+ only)
      parameters:
      - description: Role details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCustomRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CustomRoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create custom role
      tags:
      - custom-fields
  /api/custom-fields/roles/{role_id}:
    delete:
      consumes:
      - application/json
      description: Delete a custom role (ADMIN+ only, system defaults cannot be deleted)
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete custom role
      tags:
      - custom-fields
    get:
      consumes:
      - application/json
      description: Get a specific custom role by ID
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CustomRoleResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get custom role
      tags:
      - custom-fields
    put:
      consumes:
      - application/json
      description: Update a custom role (ADMIN+ only, system defaults cannot be updated)
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      - description: Role updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCustomRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CustomRoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update custom role
      tags:
      - custom-fields
  /api/custom-fields/stages:
    post:
      consumes:
      - application/json
      description: Create a new custom stage for a project (ADMIN+ only)
      parameters:
      - description: Stage details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCustomStageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CustomStageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create custom stage
      tags:
      - custom-fields
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"